// src/app/Models/dashboard.ts
export class WidgetData {
  count: string;
  message: string;
  status: string;
  title: string;
}
export class BUsAndJobRequestCount {
  buId: string;
  buName: string;
  count: string;
  message: string;
  status: 0;
}

export class LatestJobRequests {
  buId: string;
  buName: string;
  closedJobRequest: string;
  closedOpenings: string;
  contractDuration: string;
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  currency: string;
  customerId: string;
  customerName: string;
  daysLeft: string;
  daysSpent: string;
  daysTotal: string;
  employmentType: string;
  holdJobRequest: string;
  id: string;
  jobDescription: string;
  jobReqStatus: string;
  location: string;
  mandatorySkills: string;
  message: string;
  monthOrYear: string;
  noOfOpenings: string;
  openJobRequest: string;
  optionalSkills: string;
  payFrequency: string;
  payRange: string;
  placementFor: string;
  progress: string;
  progressJobRequest: string;
  projectStartDate: string;
  recruiterId: string;
  recruiterName: string;
  referenceNumber: string;
  remoteOption: string;
  requesterId: string;
  requesterName: string;
  roleId: string;
  roleName: string;
  runningNumber: string;
  status: string;
  totalJobRequest: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  year: 0;
  customerNameInDetail: string;
}

export class jobRequestCount {
  contract: number;
  crCleared: number;
  fullTime: number;
  hrCleared: number;
  ir1Cleared: number;
  ir2Cleared: number;
  message: string;
  onboarded: number;
  resumesShortlisted: number;
  resumesTagged: number;
  selected: number;
  status: number;
  totalRequest: 0;
}